feat: broken file upload
Change Counter island to a form for file upload.
The form has submit handler to `PUT` the file with `fetch`.
There is a route handler for `PUT` that redirects to a testing API.

At the moment, submitting the form results in `ERR_CONNECTION_RESET`.

If `fetch` with other `body` content, it works fine.